Move switch and indicator to design system
