+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/screen/NotificationSettingsScreen.kt
@@ -20,7 +20,7 @@ import net.mullvad.mullvadvpn.compose.preview.NotificationSettingsUiStatePreview
 import net.mullvad.mullvadvpn.compose.state.NotificationSettingsUiState
 import net.mullvad.mullvadvpn.compose.transitions.SlideInFromRightTransition
 import net.mullvad.mullvadvpn.lib.theme.AppTheme
-import net.mullvad.mullvadvpn.lib.ui.component.MullvadCircularProgressIndicatorLarge
+import net.mullvad.mullvadvpn.lib.ui.designsystem.MullvadCircularProgressIndicatorLarge
 import net.mullvad.mullvadvpn.util.Lc
 import net.mullvad.mullvadvpn.viewmodel.NotificationSettingsViewModel
 import org.koin.androidx.compose.koinViewModel
